Refactor map pane out
