+++ org.mwc.debrief.lite/src/main/java/org/mwc/debrief/lite/map/GeoToolMapRenderer.java
@@ -14,37 +14,22 @@
  */
 package org.mwc.debrief.lite.map;
 
-import java.awt.Color;
 import java.awt.Component;
-import java.awt.Cursor;
 import java.awt.Graphics;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.geotools.data.simple.SimpleFeatureSource;
-import org.geotools.factory.Hints;
-import org.geotools.geometry.DirectPosition2D;
 import org.geotools.map.MapContent;
 import org.geotools.referencing.CRS;
-import org.geotools.referencing.crs.DefaultGeographicCRS;
 import org.geotools.renderer.lite.StreamingRenderer;
-import org.geotools.swing.JMapPane;
-import org.geotools.swing.event.MapMouseAdapter;
-import org.geotools.swing.event.MapMouseEvent;
-import org.geotools.swing.event.MapMouseListener;
-import org.geotools.swing.tool.CursorTool;
-import org.mwc.debrief.lite.DebriefLiteApp;
 import org.opengis.feature.simple.SimpleFeatureType;
-import org.opengis.geometry.MismatchedDimensionException;
 import org.opengis.referencing.FactoryException;
 import org.opengis.referencing.crs.CoordinateReferenceSystem;
 import org.opengis.referencing.operation.MathTransform;
-import org.opengis.referencing.operation.TransformException;
 
 import Debrief.GUI.Frames.Application;
 import MWC.GUI.ToolParent;
-import MWC.GenericData.WorldLocation;
-import MWC.Utilities.TextFormatting.BriefFormatLocation;
 
 /**
  *
@@ -54,167 +39,12 @@ import MWC.Utilities.TextFormatting.BriefFormatLocation;
 public class GeoToolMapRenderer
 {
 
-  private static class CustomMapPane extends JMapPane
-  {
-
-    /**
-     *
-     */
-    private static final long serialVersionUID = 1L;
-
-    private static final String WORLD_PROJECTION = "EPSG:3395"; // 3395 for Mercator proj (? or may
-                                                                // be 3857?)
-    private static final String DATA_PROJECTION = "EPSG:4326";
-    private final MouseDragLine dragLine;
-
-    private final GeoToolMapRenderer _renderer;
-    private final MapMouseListener mouseMotionListener = new MapMouseAdapter()
-    {
-
-      void handleMouseMovement(final MapMouseEvent ev)
-      {
-        // mouse pos in Map coordinates
-        final DirectPosition2D curPos = ev.getWorldPos();
-
-        if (ev.getWorldPos()
-            .getCoordinateReferenceSystem() != DefaultGeographicCRS.WGS84)
-        {
-          try
-          {
-            data_transform.transform(curPos, curPos);
-          }
-          catch (MismatchedDimensionException | TransformException e)
-          {
-            Application.logError2(ToolParent.ERROR,
-                "Failure in projection transform", e);
-          }
-        }
-
-        final WorldLocation current = new WorldLocation(curPos.getY(), curPos
-            .getX(), 0);
-        final String message = BriefFormatLocation.toString(current);
-        DebriefLiteApp.updateStatusMessage(message);
-      }
-
-      @Override
-      public void onMouseDragged(final MapMouseEvent arg0)
-      {
-        if (!(currentCursorTool instanceof RangeBearingTool))
-        {
-          handleMouseMovement(arg0);
-        }
-      }
-
-      @Override
-      public void onMouseEntered(final MapMouseEvent arg0)
-      {
-        handleMouseMovement(arg0);
-      }
-
-      @Override
-      public void onMouseExited(final MapMouseEvent arg0)
-      {
-        handleMouseMovement(arg0);
-      }
-
-      @Override
-      public void onMouseMoved(final MapMouseEvent arg0)
-      {
-        handleMouseMovement(arg0);
-      }
-
-      @Override
-      public void onMouseWheelMoved(final MapMouseEvent arg0)
-      {
-        handleMouseMovement(arg0);
-      }
-    };
-
-    private CoordinateReferenceSystem worldCoords;
-
-    private CoordinateReferenceSystem worldDegs;
-
-    private MathTransform data_transform;
-
-    public CustomMapPane(final GeoToolMapRenderer geoToolMapRenderer)
-    {
-      super();
-      
-      // Would be better to pass in a GeoToolMapProjection or GTProjection here?
-      try
-      {
-        worldCoords = CRS.decode(WORLD_PROJECTION);
-
-        Hints.putSystemDefault(Hints.FORCE_LONGITUDE_FIRST_AXIS_ORDER,
-            Boolean.TRUE);
-        worldDegs = CRS.decode(DATA_PROJECTION);
-        data_transform = CRS.findMathTransform(worldCoords, worldDegs);
-      }
-      catch (final FactoryException e)
-      {
-        Application.logError2(ToolParent.ERROR,
-            "Failure in projection transform", e);
-      }
-      _renderer = geoToolMapRenderer;
-
-      dragLine = new MouseDragLine(this);
-      addMouseListener(dragLine);
-      addMouseMotionListener(dragLine);
-      addMouseListener(mouseMotionListener);
-      
-      // try to set background color
-      super.setBackground(new Color(135, 172, 215));
-    }
-
-    @Override
-    protected void paintComponent(final Graphics arg0)
-    {
-      super.paintComponent(arg0);
-      _renderer.paintEvent(arg0);
-    }
-
-    @Override
-    public void setCursorTool(final CursorTool tool)
-    {
-      paramsLock.writeLock().lock();
-      try
-      {
-        if (currentCursorTool != null)
-        {
-          mouseEventDispatcher.removeMouseListener(currentCursorTool);
-        }
-
-        currentCursorTool = tool;
-
-        if (currentCursorTool == null)
-        {
-          setCursor(Cursor.getDefaultCursor());
-          dragBox.setEnabled(false);
-          dragLine.setEnabled(false);
-        }
-        else
-        {
-          setCursor(currentCursorTool.getCursor());
-          dragLine.setEnabled(currentCursorTool instanceof RangeBearingTool);
-          dragBox.setEnabled(currentCursorTool.drawDragBox());
-          currentCursorTool.setMapPane(this);
-          mouseEventDispatcher.addMouseListener(currentCursorTool);
-        }
-
-      }
-      finally
-      {
-        paramsLock.writeLock().unlock();
-      }
-    }
-  }
-
   public static interface MapRenderer
   {
     public void paint(final Graphics gc);
   }
 
-  private CustomMapPane mapPane;
+  private LiteMapPane mapPane;
 
   private final MapContent mapContent;
 
@@ -238,12 +68,13 @@ public class GeoToolMapRenderer
     _myRenderers.add(renderer);
   }
 
-  public void createMapLayout()
+  public LiteMapPane createMapLayout()
   {
-    mapPane = new CustomMapPane(this);
+    mapPane = new LiteMapPane(this);
     final StreamingRenderer streamer = new StreamingRenderer();
     mapPane.setRenderer(streamer);
     mapPane.setMapContent(mapContent);
+    return mapPane;
   }
 
   /**
@@ -301,7 +132,7 @@ public class GeoToolMapRenderer
     return transform;
   }
 
-  private void paintEvent(final Graphics arg0)
+  void paintEvent(final Graphics arg0)
   {
     for (final MapRenderer r : _myRenderers)
     {
